Do not call Dispose() as ClassicPendingScript's prefinalizer
Because the things in ClassicPendingScript::DisposeInternal() except for
ScriptStreamer::Cancel() doesn't have to be called as a prefinalizer,
this CL introduces ClassicPendingScript::Prefinalize() that only calls
ScriptStreamer::Cancel() and thus makes Dispose() not to be called there.

This CL simplified the prefinalization of ClassicPendingScript, especially
order dependencies between ClassicPendingScript's prefinalizer and the
prefinalizer of its parent class (ResourceOwner).
Leaving ClassicPendingScript in a not-Dispose()d state is not observable
if the related classes obeys the rule of Oilpan, and
https://codereview.chromium.org/2837413002/ checks that in case there were
a bug.

BUG=715309

Review-Url: https://codereview.chromium.org/2844583002
Cr-Original-Commit-Position: refs/heads/master@{#467299}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: faa055103d806be6bcdb05b721c954d1b1584200