Worklet: Move Worklet::Initialize() into ThreadedWorklet for cleanup
This is split from https://codereview.chromium.org/2840523002/

Worklet::Initialize() is called when addModule() is called for the first time,
and expected to initialize a messaging proxy and a worklet thread. This is
used only for threaded worklets, and not necessary for main thread worklets.
Therefore, this CL moved it from Worklet to ThreadedWorklet.

Motivation of this change: MainThreadWorklet will have a separate initialization
logic after the above CL and the similar initialization function will be an
obstacle.

BUG=627945

Review-Url: https://codereview.chromium.org/2845533002
Cr-Original-Commit-Position: refs/heads/master@{#467300}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f2c7bd6ebff851528ab06aa8ff14a964e3437df0