Use dark scrollbar theme for transparent frames.
If the given background color is totally transparent we should default
to using the dark theme since that's the default. ChromeOS OOBE pages,
for example, set an override background color of "TRANSPARENT" which is
#00000000 in RGBA. Our old logic looked only at the RGB bits so it
concluded that the frame had a black background and used white scollbars.

We don't have enough information in Blink to know what theme to use in
this case but falling back to the default theme should be good enough.

BUG=709978

Review-Url: https://codereview.chromium.org/2840223002
Cr-Original-Commit-Position: refs/heads/master@{#467382}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f82b1ff2761897b38ae79042810fb0a9f3ff747f