Add automated VR controller emulator
Adds the Daydream controller test API, includes it in the test APK,
and adds a simple WebVR test using it. The wrapper around the API
will be further improved in the future once VR Shell testing becomes
the focus - for now, it only supports clicking of the touchpad button.

BUG=713342

Review-Url: https://codereview.chromium.org/2826273002
Cr-Original-Commit-Position: refs/heads/master@{#467420}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 62d159fb42542807142a39033744c73bbb506698