LayoutTableCell::OffsetFromContainer() should use the flipped offset of its parent.
Since LayoutTableCell is relative to LayoutTableSection, it should subtract the
offset of the LayoutTableRow. OffsetFromContainer() is flipped value, but
LayoutTableCell::OffsetFromContainer() uses LocationOffset() which is an unflipped
location. Therefore, the rect of td goes outside of tr.

It should use PhysicalLocationOffset() which returns a flipped location.

BUG=711193

Review-Url: https://codereview.chromium.org/2816983002
Cr-Original-Commit-Position: refs/heads/master@{#467467}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5cdd84c272ce41e804a9d4447b288f44d748c6af