Do not count extra trailing space for the last word in a range
Sometimes when a text selection ends with the last word on a line
(for instance when just one word is selected) the generated text
for the selection range contains an extra trailing space.
This happens when a text node is split into several InlineTextBoxes
(e.g. a <span> inside a narrow table) and the last selected word
happens to be the last one in such text box.

The TextIterator algorithm restores the spaces between the text boxes
and will add a trailing space for this word. For the SurroundingText,
however, we need the length that corresponds to the visible selection.

This CL introduces a new text iterator behavior that prevents that
extra space to be added. This behavior is used in combination with
DefaultRangeLengthBehavior() for the selection range calculation
in the SurroundingText.

BUG=708329

Review-Url: https://codereview.chromium.org/2832123002
Cr-Original-Commit-Position: refs/heads/master@{#467538}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9305554380ee3a02c680009b5dee4b9e82cba334