Store nonAttachedStyle on NodeLayoutData instead of on HashMap in Document.
This patch stores the ComputedStyle to be passed between recalcStyle
and rebuildLayoutTree on NodeLayoutData instead of storing it on the
HashMap on Document.

This potentially fixes the perf regression in crbug.com/700093 as it
removes the HashMap lookup in SharedStyleFinder and making it a free
operation again.

BUG=595137, 700093

Review-Url: https://codereview.chromium.org/2821193003
Cr-Original-Commit-Position: refs/heads/master@{#467542}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 95862a52c7bc1a5846259619243289805d095b70