Do not convert mouse move events to mouse leave events in blink
In blink, we convert mouse enter and leave events to mouse move events,
we should not do it anymore. We should keep their types, so we know
they are actual enter or leave events.

I looked at the code, PointerEventManager::SendMousePointerEvent has a
separate argument called mouse_event_type, which is always the type of
mouse move, and mouseevent's type is not referred anymore. It should
be safe to remove the conversion.

BUG=686196

Review-Url: https://codereview.chromium.org/2839103002
Cr-Original-Commit-Position: refs/heads/master@{#467674}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b40772db092d28a4abfa3b0d869b0cb9815f3d1b