WebVR: fix focus while presenting
If the WebVR document loses focus, check if the new focused element is an
embedding local parent frame. If that's the case, continue presenting.  This
fixes issues with Cardboard-style touch events that are reported as a click at
viewport (0, 0) where the resulting focus loss stopped presentation.

BUG=710863

Review-Url: https://codereview.chromium.org/2847233002
Cr-Original-Commit-Position: refs/heads/master@{#468140}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 984c8d18acf4b891d5bf39db6a898e2187340e0f