Prune ComputeVisibleSelectionInDOMTreeDeprecated from Editor::ChangeSelectionAfterCommand
This patch prunes a call site of the deprecated function to improve
code health.

The change is safe because the call site only needs the selection type,
not the selection itself. After this patch, it also aligns with the
code in Editor::RespondToChangedSelection.

BUG=698633
TEST=n/a; no behavioral change

Review-Url: https://codereview.chromium.org/2852913002
Cr-Original-Commit-Position: refs/heads/master@{#468304}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 08b287f8807022c5c6369eafccad38ad3492cd9f