Add a doc in the presubmit script checking if layout tests use eventSender
In https://codereview.chromium.org/2810943005/, we added a presubmit script
to check if eventSender is still used in new layout tests. Here, We add a
doc to explain why we want to replace EventSender and give some examples
on how to use this new API in testing.

BUG=711340

Review-Url: https://codereview.chromium.org/2844803005
Cr-Original-Commit-Position: refs/heads/master@{#468323}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3e1724e5ccd3d41d417d13487c4b1376a6815a93