Upstream service worker `fetch` tests to WPT
Reconcile the remaining Service Worker tests that concern the Fetch
standard. As noted below, some tests must be persisted in the Chromium
project due to unresolved timeout issues in the build environment (see
https://crbug.com/658997).

**fetch-event-async-respond-with**

The Chromium-specific version of this test is almost identical to the
version provided by the Web Platform Test suite, though it suffers from
timeout issues in the build environment.

Rename the Chromium-specific version to document its deprecated status
and improve the upstream version to defer frame removal and omit
unnecessary resource.

**fetch-event-network-error**

The upstream version of this test is almost identical to the
Chromium-specific version.

The Chromium-specific version of this test is almost identical to the
version available upstream. The upstream version is preferable because
it contains an additional assertion not found in the Chromium version.

Remove the Chromium-specific version of this test, and update the
upstream version to omit an unnecessary resource.

**fetch-event-respond-with-argument**

Re-locate Chromium-specific test to the Web Platform Test suite
directory so that it may be included in that project.

**fetch-event-respond-with-stops-propagation**

The Chromium-specific version of this test is almost identical to the
version provided by the Web Platform Test suite, though it suffers from
timeout issues in the build environment.

Rename the Chromium-specific version to reflect its deprecated status,
and improve the upstream version to defer frame removal and omit
unnecessary resource.

**fetch-event**

Remove sub-tests from the Chromium-specific version that are present in
the upstream version. Rename the test file to reflect the
implementation-specific nature of the test.

Extend the upstream version with a relaxed version of Chromium's
sub-test for request headers.

In both versions, formalize tear down logic to ensure consistent frame
removal and subsequent client de-registration.

BUG=688116, 658997
R=mek@chromium.org

Review-Url: https://codereview.chromium.org/2836233002
Cr-Original-Commit-Position: refs/heads/master@{#468341}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 07b83d70ca740271a8acaa3c3e9b32eb01d5c48b