[LayoutNG] Change anonymous LayoutBlockFlow to LayoutNGBlockFlow
Additionally this resolves an incoming margin strut once we know the inline node
isn't empty. In this respect we treat it like a zero-height div.

This regresses some tests that were using the existing LayoutBlockFlow layout.
I.e. when we are inside a LayoutTableCell.

BUG=635619

Review-Url: https://codereview.chromium.org/2851983002
Cr-Original-Commit-Position: refs/heads/master@{#468470}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dc843fac8da1a46453dd6035c910c07f5237922c