[ServiceWorker] Mojofy InstallEvent of Service Worker
This CL converts install event IPCs into mojo interface:
  ServiceWorkerMsg_InstallEvent
  ServiceWorkerHostMsg_InstallEventFinished

Also add an associated interface for the ForeignFetchRegistration,
so that this event be sent on the same message pipe as the service
worker event dispatcher interface, so that we can guarantee the
delivery order of foreign fetch register calls and event dispatcher
interface callbacks.

Also add the test cases for the associated interface.

BUG=629701
TEST=content_unittests

Review-Url: https://codereview.chromium.org/2746783002
Cr-Original-Commit-Position: refs/heads/master@{#468544}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: aeae396182b48115df4170d4e6e61aea9b2e7416