Remove the last references from production code to WebViewImpl.
This CL is a continuation of removing dependencies in the code on WebViewImpl, by
introducing the pure virtual class WebViewBase and moving methods from WebViewImpl
to virtual methods in WebViewBase and overriding them in WebViewImpl.

It also removes a few of the cast to WebViewImpl, and replaces them with a direct
static cast to WebViewBase.

BUG=712963

Review-Url: https://codereview.chromium.org/2854783003
Cr-Original-Commit-Position: refs/heads/master@{#468580}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ceb538b1bb7f9067f80535bd00f49b81344b8823