[selectors4] :focus-within test when iframe loses focus
This test should be part of https://codereview.chromium.org/2821303005
but at that time I didn't find the way to focus/unfocus an iframe.
The solution was to use internals.setFocused() for that purpose.

This patch adds a new test pretty similar to the one for windows
(fast/selectors/focus-within-window-inactive.html),
but checking what happens when an iframe loses and regains focus.

BUG=617371
TEST=fast/selectors/focus-within-iframe.html

Review-Url: https://codereview.chromium.org/2847803007
Cr-Original-Commit-Position: refs/heads/master@{#468601}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 004ddc7b1a8b2c907f31179020ed17186b6639eb