Load image placeholders when the Client LoFi previews feature is on.
Design doc (public):
https://docs.google.com/document/d/1rWCON90u3j8JtgRk_SB6JTOIeKqp4rQjlRfj_W1zc3o/edit#

This CL adds a pathway between Blink and the RenderFrameImpl to
determine if both Client LoFi is enabled for a frame and Client LoFi
isn't preempted by other Previews features for a particular image.
Images for which these conditions are true are loaded as image
placeholders in order to save data.

BUG=657997

Review-Url: https://codereview.chromium.org/2784743002
Cr-Original-Commit-Position: refs/heads/master@{#469078}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 614e7167c693575e900e7e986367cd0b6d97c191