Drop GetFrame()->GetDocument() usage in FrameFetchContext
It's a workaround for the fact that FrameFetchContext::GetDocument() will be
null before FrameFetchConext::document_loader_ commits, and in that case,
it's not right to go to GetFrame()->GetDocument(), which is an unrelated
Document. Either get to the relevant ExecutionContext, or use LocalFrame
directly where an ExecutionContext/SecurityContext isn't needed (e.g., for UseCounters).

BUG=

Review-Url: https://codereview.chromium.org/2844913002
Cr-Original-Commit-Position: refs/heads/master@{#469127}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 20f06e18b747e0f99ed657ce461c4db97fb97f7c