Refactor prepareToMeasureValuesAsync to startMeasureValuesAsyn which run test through callback
Motivation:
In order to support tracing metrics for prepareToMeasureValuesAsync method,
we need to be able to trigger tracing between prepareToMeasureValuesAsync &
actual test run. This CL rename prepareToMeasureValuesAsync method to
startMeasureValuesAsyn & refactor it  so that the
actual test code is specified as a callback, which allows the framework
to interject the startTrace() call before running test in the future.

BUG=701059

Review-Url: https://codereview.chromium.org/2858783003
Cr-Original-Commit-Position: refs/heads/master@{#469233}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8dc2204a41db0b7b0f66c529c53d4c4a5d9d7fb5